// src/app/modules/dashboard/dashboard.interface.ts
export interface DashboardQuery {
  incomeYear?: number;
  joinYear?: number;
  role?: string;
}

export interface MonthlyIncome {
  month: string;
  income: number;
}

export interface MonthlyOrders {
  month: string;
  total: number;
}
export interface MonthlyUsers {
  month: string;
  total: number;
}
export interface monthlyOrders {
  month: string;
  total: number;
}

export interface DashboardData {
  totalUsers: number;
  totalPayout: number;
  totalCustomers: number;
  totalDriver: number;
  totalIncome: number;
  toDayIncome: number;
  monthlyIncome: MonthlyIncome[];
  monthlyUsers: MonthlyUsers[];
  monthlyOrders: monthlyOrders[];
}